services/writeops/requirements.txt
# Write Operations Service Dependencies

# Core framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0

# Environment configuration
python-dotenv==1.0.0
